requirements.txt
XlsxWriter
pickleshare
pyobjc-framework-SystemConfiguration
imutils
numpy
matplotlib
Pillow
